Cset exclude: vvaintroub@wva.|ChangeSet|20080510002431|10233

+++ storage/falcon/Cache.cpp
@@ -818,7 +818,7 @@ void Cache::ioThread(void)
 						
 					flushLock.unlock();
 					//Log::debug(" %d Writing %s %d pages: %d - %d\n", thread->threadId, (const char*) dbb->fileName, count, pageNumber, pageNumber + count - 1);
-					int length = (int)(p - buffer);
+					int length = p - buffer;
 					priority.schedule(PRIORITY_LOW);
 					
 					try
